Fix Wi-Fi access point configuration with RTL8733BU (#146)

+++ configs/usr/lib/wb-configs/fix_nm_ap_ssid.sh
@@ -2,6 +2,34 @@
 
 wb_ap_ssid_prefix="${WB_AP_SSID_PREFIX:-WirenBoard}"
 
+# find Wi-Fi module
+wifi_module="RTL8723BU"
+if [ -z $(lsusb -d'0bda:b720') ]; then
+    wifi_module="RTL8733BU"
+fi
+
+# Args:
+# - MD5 hash of wb-ap.nmconnection without ssid and timestamp
+# - Wi-Fi module name
+should_delete_config() {
+    # remove old wb-ap without subnet address
+    if [ "$1" == "2ed8e74107f2dc68fabe96cd0fe6b61c" ]; then
+        return 0
+    fi
+
+    # remove old wb-ap without channel
+    if [ "$1" == "f2a720dc98d3764efab777f060ca79c4" ]; then
+        return 0
+    fi
+
+    # remove old wb-ap with band and channels
+    if [ "$2" == "RTL8733BU" ] && [ "$1" == "1737c3e9aa4a887f59235c6b4867fa5d" ]; then
+        return 0
+    fi
+
+    return 1
+}
+
 # remove old wi-fi ap connection
 wb_ap_connection="/etc/NetworkManager/system-connections/wb-ap.nmconnection"
 short_sn=`wb-gen-serial -s`
@@ -13,18 +41,15 @@ if [ -f $wb_ap_connection ]; then
     # remove timestamp
     sed -i "/timestamp/d" $tmp_connection
     MD5=$(md5sum "$tmp_connection" | cut -f 1 -d ' ')
-    # remove old wb-ap without subnet address
-    if [ "$MD5" == "2ed8e74107f2dc68fabe96cd0fe6b61c" ]; then
-        rm -f $wb_ap_connection || true
-    fi
 
-    # remove old wb-ap without channel
-    if [ "$MD5" == "f2a720dc98d3764efab777f060ca79c4" ]; then
+    if should_delete_config $MD5 $wifi_module; then
         rm -f $wb_ap_connection || true
     fi
 
     # Bad generated ssid
-    grep -q -x -F "ssid=WirenBoard-" "$wb_ap_connection" && rm -f $wb_ap_connection
+    if [ -f $wb_ap_connection ]; then
+        grep -q -x -F "ssid=WirenBoard-" "$wb_ap_connection" && rm -f $wb_ap_connection
+    fi
 fi
 
 wb_ap_connection_template="/usr/lib/NetworkManager/system-connections/wb-ap.nmconnection"
@@ -35,4 +60,10 @@ if [ ! -f "$wb_ap_connection" ] && [ ! -f "$deleted_wb_ap" ]; then
     ssid="${wb_ap_ssid_prefix}-${short_sn}"
     cp -f $wb_ap_connection_template $wb_ap_connection
     sed -i "s/^ssid=@SSID@$/ssid=${ssid}/" "$wb_ap_connection"
+
+    # remove band and channel for new chipset
+    if [ "$wifi_module" == "RTL8733BU" ]; then
+        sed -i "/channel=1/d" "$wb_ap_connection"
+        sed -i "/band=bg/d" "$wb_ap_connection"
+    fi
 fi
